refactor: 선택된 쿠폰 전역 상태 구조 롤백
